project struct adjust
