+++ controller/similarTest.go
@@ -1,0 +1,12 @@
+package controller
+
+import(
+	// "net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func SimilarTest(c *gin.Context) {
+	
+
+}
