+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/RaymondCode/simple-demo
+module SimpleDouyin
 
 go 1.17
 
@@ -52,6 +52,6 @@ require (
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 	gorm.io/driver/mysql v1.5.1
-	gorm.io/gorm v1.25.2 // indirect
+	gorm.io/gorm v1.25.2
 	moul.io/http2curl/v2 v2.3.0 // indirect
 )
