+++ router/router.go
@@ -1,0 +1,48 @@
+package main
+
+import (
+	"net/http"
+
+	"SimpleDouyin/controller"
+	"SimpleDouyin/repository"
+	"github.com/gin-gonic/gin"
+)
+
+func initRouter(r *gin.Engine) {
+	// public directory is used to serve static resources
+	r.Static("/static", "./public")
+	r.LoadHTMLGlob("templates/*")
+
+	// home page
+	r.GET("/", func(c *gin.Context) {
+		c.HTML(http.StatusOK, "index.html", gin.H{
+			"title": "Main website",
+		})
+	})
+
+	apiRouter := r.Group("/douyin")
+
+	// basic apis
+	apiRouter.GET("/LarryTest/", repository.LarryTest)
+	apiRouter.GET("/SimilarTest/",controller.SimilarTest)
+	apiRouter.GET("/feed/", controller.Feed)
+	apiRouter.GET("/user/", controller.UserInfo)
+	apiRouter.POST("/user/register/", controller.Register)
+	apiRouter.POST("/user/login/", controller.Login)
+	apiRouter.POST("/publish/action/", controller.Publish)
+	apiRouter.GET("/publish/list/", controller.PublishList)
+
+	// extra apis - I
+	apiRouter.POST("/favorite/action/", controller.FavoriteAction)
+	apiRouter.GET("/favorite/list/", controller.FavoriteList)
+	apiRouter.POST("/comment/action/", controller.CommentAction)
+	apiRouter.GET("/comment/list/", controller.CommentList)
+
+	// extra apis - II
+	apiRouter.POST("/relation/action/", controller.RelationAction)
+	apiRouter.GET("/relation/follow/list/", controller.FollowList)
+	apiRouter.GET("/relation/follower/list/", controller.FollowerList)
+	apiRouter.GET("/relation/friend/list/", controller.FriendList)
+	apiRouter.GET("/message/chat/", controller.MessageChat)
+	apiRouter.POST("/message/action/", controller.MessageAction)
+}
